fix(frontend): CodeRabbit review — shared useBoardSelection, tile toggle out of the Link
Extract the duplicated board-selection-from-URL derivation (route match +
home lens + six-axis parse) into hooks/use-board-selection.ts, consumed by
both BoardModeBlock and BoardFilterChips so the URL vocabulary has one
reader (INV-35). Move BoardTileToggle out of the row <Link> — a button may
not nest inside an anchor — rendering it as an absolutely positioned
sibling at the same visual spot (strip-aware left offset), in both
StreamItem and ScratchpadItem; StreamItemAvatar loses the boardControl slot.
